geo bounding tests
