+++ src/Nest.Searchify/Queries/Models/GeoBoundingBox.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Linq;
 using System.Text.RegularExpressions;
+using Newtonsoft.Json;
 
 namespace Nest.Searchify.Queries.Models
 {
