+++ tests/Nest.Searchify.Tests/Nest.Searchify.Tests.csproj
@@ -36,6 +36,7 @@
   <ItemGroup>
     <Compile Include="Extensions\ReadEmbeddedResourceExtensions.cs" />
     <Compile Include="FilterFieldTests\FilterFieldContext.cs" />
+    <Compile Include="GeoPointTests\GeoBoundingBoxTests.cs" />
     <Compile Include="GeoPointTests\GeoPointContext.cs" />
     <Compile Include="Integration\InMemoryElasticsearchContext.cs" />
     <Compile Include="Integration\Person.cs" />
